pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.1",
    "ipykernel>=7.1.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.1",
    "langchain-openai>=1.1.1",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
